# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/{userId}/legal-representative": {
      "post": {
        "operationId": "LegalRepresentativesController_create_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLegalRepresentativeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Representante registrado"
          },
          "400": {
            "description": "Usuario mayor de edad o ya tiene representante"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar representante legal (solo para menores de 18)",
        "tags": [
          "legal-representatives"
        ]
      },
      "get": {
        "operationId": "LegalRepresentativesController_findOne_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta exitosa. Retorna representante o data=null cuando no existe"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener representante legal del usuario",
        "tags": [
          "legal-representatives"
        ]
      },
      "patch": {
        "operationId": "LegalRepresentativesController_update_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLegalRepresentativeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Representante actualizado"
          },
          "404": {
            "description": "Representante no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar representante legal",
        "tags": [
          "legal-representatives"
        ]
      },
      "delete": {
        "operationId": "LegalRepresentativesController_remove_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Representante eliminado"
          },
          "404": {
            "description": "Representante no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar representante legal",
        "tags": [
          "legal-representatives"
        ]
      }
    },
...
    "/v1/honors": {
      "get": {
        "description": "Lista todos los honores activos con paginación y filtros",
        "operationId": "HonorsController_findAll_v1",
        "parameters": [
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "clubTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skillLevel",
            "required": false,
            "in": "query",
            "description": "1=Básico, 2=Avanzado, 3=Máster",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de honores"
          }
        },
        "summary": "Listar honores",
        "tags": [
          "honors"
        ]
      }
    },
...
    "/v1/users/{userId}/honors": {
      "get": {
        "description": "Lista los honores en los que el usuario está inscrito o ha completado",
        "operationId": "UserHonorsController_getUserHonors_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "validated",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Honores del usuario"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener honores del usuario",
        "tags": [
          "user-honors"
        ]
      },
      "post": {
        "description": "Registra (o reactiva) un honor del usuario permitiendo enviar evidencias en el alta",
        "operationId": "UserHonorsController_createUserHonor_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserHonorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Honor registrado para el usuario"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar honor con datos iniciales",
        "tags": [
          "user-honors"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateLegalRepresentativeDto": {
        "type": "object",
        "properties": {
          "representative_user_id": {
            "type": "string",
            "description": "ID del usuario registrado que es el representante legal"
          },
          "name": {
            "type": "string",
            "description": "Nombre (requerido si no se proporciona representative_user_id)"
          },
          "paternal_last_name": {
            "type": "string",
            "description": "Apellido paterno (requerido si no se proporciona representative_user_id)"
          },
          "maternal_last_name": {
            "type": "string",
            "description": "Apellido materno"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono (requerido si no se proporciona representative_user_id)"
          },
          "relationship_type_id": {
            "type": "string",
            "description": "UUID del tipo de relación"
          }
        },
        "required": [
          "relationship_type_id"
        ]
      },
      "UpdateLegalRepresentativeDto": {
        "type": "object",
        "properties": {
          "representative_user_id": {
            "type": "string",
            "description": "ID del usuario registrado que es el representante legal"
          },
          "name": {
            "type": "string",
            "description": "Nombre (requerido si no se proporciona representative_user_id)"
          },
          "paternal_last_name": {
            "type": "string",
            "description": "Apellido paterno (requerido si no se proporciona representative_user_id)"
          },
          "maternal_last_name": {
            "type": "string",
            "description": "Apellido materno"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono (requerido si no se proporciona representative_user_id)"
          },
          "relationship_type_id": {
            "type": "string",
            "description": "UUID del tipo de relación"
          }
        }
      },
...
      "CreateUserHonorDto": {
        "type": "object",
        "properties": {
          "honorId": {
            "type": "number",
            "description": "ID del honor/especialidad"
          },
          "date": {
            "type": "string",
            "description": "Fecha de registro/inicio del honor"
          },
          "validate": {
            "type": "boolean",
            "description": "Honor validado por instructor"
          },
          "certificate": {
            "type": "object",
            "description": "URL del certificado (pasar null o string vacío para limpiar)"
          },
          "images": {
            "description": "URLs de imágenes de evidencia",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "document": {
            "type": "object",
            "description": "URL del documento adicional"
          }
        },
        "required": [
          "honorId"
        ]
      },
...
    }
  }
}